requirements.txt
PyQt5
pyserial
adafruit-circuitpython-motorkit
adafruit-circuitpython-motor
platformdirs